requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.5.7
certifi==2021.10.8
cffi==1.15.0
coverage==6.3.1
cryptography==36.0.2
defusedxml==0.7.1
dj-database-url==0.4.2
django-cors-headers==3.10.1
# -e git+https://github.com/jazzband/django-debug-toolbar.git@c7123afdf495c03309e38c57674571cb02165de0#egg=django_debug_toolbar
django-environ==0.8.0
django-filter==2.4.0
django-rest-authtoken==2.1.3
djangorestframework==3.13.0
djangorestframework-xml==2.0.0
drf-nested-routers==0.93.4 
flake8==3.9.2
gunicorn==20.0.4
h11==0.13.0
idna==3.3
mccabe==0.6.1
# nltk==3.7
outcome==1.1.0
# pandas==1.1.5 
psycopg2==2.9.3
psycopg2-binary==2.9.2
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
pyOpenSSL==22.0.0
PySocks==1.7.1
selenium==4.1.2
# numpy==1.19.5
# scipy==1.4.1
# scikit-learn==0.22.2.post1
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
whitenoise==5.3.0
wsproto==1.1.0
#New
amqp==5.1.1
billiard<4.0,>=3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-repl==0.2.0
Django==4.0.4
kombu == 5.2.4
prompt-toolkit==3.0.29
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
django-celery-results==2.3.1
flower == 1.0.0
pytest ==7.1.2
pytest-django == 4.5.2
